Made code more DRY. Put handleSubmit, handleChange, and error display into helper functions.
