+++ client/src/views/forms/components/ClassForm.js
@@ -1,6 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import {classFormSchema} from '../validation/schema';
-import {validateForm,validateField} from '../validation/validationHelpers';
+import {validateForm} from '../validation/validationHelpers';
+import {displayErrors,handleChangeHelper,handleSubmitHelper} from '../formHelpers';
 const initialValues = {
     className:'',
     classType:'',
@@ -24,16 +25,18 @@ function ClassForm(){
     },[]);
     // function declarations
     const handleChange=(event)=>{
-        const {name,value,checked,type} = event.target;
-        const inputValue = type==='checkbox'?checked:value;
-        const newFormValues = {...formValues,[name]:inputValue};
-        validateField(classFormSchema,name,inputValue,formErrors,setFormErrors); //validate changed field using yup.reach
-        validateForm(classFormSchema,newFormValues,setIsValid);
-        setFormValues(newFormValues);
+        handleChangeHelper({
+            event,
+            schema:classFormSchema,
+            formValues,
+            setFormValues,
+            formErrors,
+            setFormErrors,
+            setIsValid
+        });
     };
     const handleSubmit=(event)=>{
-        event.preventDefault();
-        // redux refactoring
+        handleSubmitHelper(event);
     };
     return(
         <form onSubmit={handleSubmit}>
@@ -75,9 +78,7 @@ function ClassForm(){
                 <input type='number' name='maxClassSize' value={formValues.maxClassSize} onChange={handleChange}></input>
             </label>
             <button type='submit' disabled={!isValid}>Add/Edit Class</button>
-            {
-                Object.keys(formErrors).map((key,i)=>formErrors[key]===''?'':<div key={i}>{formErrors[key]}</div>)
-            }
+            {displayErrors(formErrors)}
         </form>
     );
 }
